+++ packages/react/src/types/client.ts
@@ -1,4 +1,4 @@
-import { MediaSdkClient } from '@cere-media-sdk/client';
+import { MediaSdkClient } from '@cere/media-sdk-client';
 
 export interface UseMediaClientReturn {
   client?: MediaSdkClient;