+++ playground/src/components/playground/Playground.tsx
@@ -1,4 +1,4 @@
-import { MediaSdkClientProvider } from '@cere-media-sdk/react';
+import { MediaSdkClientProvider } from '@cere/media-sdk-react';
 import { Box } from '@mui/material';
 import { ConnectWallet, useSigner } from '@thirdweb-dev/react';
 import { PlaygroundNavigation } from './PlaygroundNavigation';